\name{EstimWHD-package}
\alias{EstimWHD-package}
\alias{EstimWHD}
\docType{package}
\title{
\packageTitle{EstimWHD}
}
\description{
\packageDescription{EstimWHD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EstimWHD}
\packageIndices{EstimWHD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EstimWHD}

Maintainer: \packageMaintainer{EstimWHD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
